Split Mux into different files for egress and ingress
